requirements.txt
hydra-core==1.3.0
pytorch-lightning==1.8.4
timm==0.6.12
